refactor: move dp metadata codec logic out of p2p, reuse it in camp2p
